fix(elevenlabs): move generateCueFile to node-only module
- Touches: lib/services/elevenlabs-service.ts (removed generateCueFile), lib/services/elevenlabs-node.ts (new), scripts/generate-cue-audio.ts, tests/unit/services/elevenlabs-service.test.ts, tests/integration/voice/{cue-generation,elevenlabs-integration}.test.ts
- Outcome: unblocks iOS Xcode build that was failing with "main.jsbundle: Invalid expression encountered" inside Apple's App Intents metadata extractor. Root cause: elevenlabs-service.ts contained `await import('node:fs')` which Metro's static analyzer bundled into main.jsbundle; the extractor then choked on the transpiled `yield import('node:fs')` syntax.
- Fix: split the Node-only filesystem helper (generateCueFile) into a new lib/services/elevenlabs-node.ts that the RN bundler never reaches. The RN-facing elevenlabs-service.ts now only exports generateSpeech / streamSpeech. Dev scripts + tests import the helper from the node module.
- Notes: the `/* webpackIgnore: true */` comment in the previous code was a webpack-only directive that Metro ignores. Updated the unit test that was asserting `result === false` (expected when the dynamic import failed) to assert `result === true` now that the helper reliably writes.
